fix(controller): keep a node's enabled flag and address across a deferred build
Four findings from the review of #869:

A peer matter.js has as disabled never connects and so never seeds, so the
reconciliation of the enabled flag - ioBroker's copy wins over the one
matter.js persisted - sat behind a gate it could never pass. `setEnabled`
did not help either, because the flag it compares still held its default.
The reconciliation now runs before the gate.

The build deferred until a peer seeds dropped the details it was called
with, and a peer restored at startup is connected before the controller
observes it, so its connection state never changes and no address is
handed over. The address is derived from the peer when the caller has
none, and `operationalAddressOf` moved next to the state it feeds.

Removing a node dropped whatever `#nodes` held rather than the object it
was tearing down, which leaks the subscriptions of a rebuild that replaced
it meanwhile. That rebuild is now also refused once the peer is gone,
instead of recreating the objects the removal deleted.
